+++ src/test/php/lang/ast/unittest/emit/TernaryTest.class.php
@@ -2,7 +2,7 @@
 
 class TernaryTest extends EmittingTest {
 
-  #[@test, @values(map= [true => "OK", false => "Fail"])]
+  #[@test, @values(['map' => [true => 'OK', false => 'Fail']])]
   public function ternary($value, $result) {
     $this->assertEquals($result, $this->run(
       'class <T> {
@@ -14,7 +14,19 @@ class TernaryTest extends EmittingTest {
     ));
   }
 
-  #[@test, @values(map= ["OK" => "OK", null => "Fail"])]
+  #[@test, @values(['map' => [true => MODIFIER_PUBLIC, false => MODIFIER_PRIVATE]])]
+  public function ternary_constants_goto_label_ambiguity($value, $result) {
+    $this->assertEquals($result, $this->run(
+      'class <T> {
+        public function run($value) {
+          return $value ?  MODIFIER_PUBLIC : MODIFIER_PRIVATE;
+        }
+      }',
+      $value
+    ));
+  }
+
+  #[@test, @values(['map' => ['OK' => 'OK', null => 'Fail']])]
   public function short_ternary($value, $result) {
     $this->assertEquals($result, $this->run(
       'class <T> {